Move WrapWithTag to VS Contracts
