+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/WrapWithTag/IWrapWithTagEndpoint.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using OmniSharp.Extensions.JsonRpc;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts.WrapWithTag;
+
+[Parallel, Method(LanguageServerConstants.RazorWrapWithTagEndpoint)]
+internal interface IWrapWithTagEndpoint : IJsonRpcRequestHandler<WrapWithTagParams, WrapWithTagResponse?>
+{
+}
